sig-cluster-lifecycle: remove ownership of kube-up
kube-up that resides in the root of the kubernetes/kubernetes
/cluster directory was originally owned by SIG Cluster Lifecycle.

In Nov 2019 the last remaining approver (roberthbailey) from the SIG
moved them self to emeritus in this OWNERS file:
https://github.com/kubernetes/kubernetes/commit/9d211ab352f7

kube-up is nowadays still heavily used in Kubernetes e2e testing,
but is unofficially maintained by SIG Testing and SIG Cloud Provider
(GCP).

Remove kube-up from the list of sub-projects of SIG Cluster Lifecycle
and leave it to another SIG if they want to take ownership of kube-up.